feat: add repl
- Built and added a new interactive REPL behind `+repl` (or `+R`).
- Implemented incremental compilation via `cc_seeded` so functions and globals persist across inputs.
- Added REPL commands:
  - `:q`/`:quit` exit
  - `:dis` disassemble last submission
  - `:stats` show bytecode stats
  - `:mem` show VM memory usage
  - `:reset` reset VM/allocator state
- Fixed printf format warnings for `int64_t`/`size_t` and adjusted `@len` to support objects.

How to run:
- Build: `make`
- Start REPL: `./build/purple_garden +repl`
- Try:
  - `@function hi[name] (@println "hi" name)`
  - `hi "pg"`
  - `:dis`, `:stats`, `:mem`, `:reset`

+++ common.c
@@ -83,7 +83,7 @@ void Value_debug(const Value *v) {
     printf("(%g)", v->floating);
     break;
   case V_INT:
-    printf("(%ld)", v->integer);
+    printf("(%lld)", (long long)v->integer);
     break;
   case V_OBJ:
     // TODO: V_OBJ
